requirements.txt
# Core Requirements
streamlit==1.31.1
opencv-python==4.9.0.80
ultralytics==8.1.27
python-dotenv==1.0.1
pydantic-ai==0.0.22 
numpy==1.26.4

# Groq API Requirements
groq==0.1.0  # If using official Groq package
httpx==0.27.0
websockets==12.0

# Async Support
asyncio==3.4.3
aiohttp==3.9.3
